requirements.txt
python-dotenv
langchain
langchain_openai
langchain-mistralai
chromadb
langchain_chroma
streamlit